requirements.txt
streamlit==1.46.1
pdfplumber==0.9.0
python-docx==0.8.11
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.3.0
